pyproject.toml
[tool.poetry]
name = "climate-risk-map-api"
version = "0.1.0"
description = "API for accessing UW CRL Lab Data"
authors = ["eric <ecol07@uw.edu>"]
license = "CC-BY-SA-4.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.10"
fastapi = "^0.115.3"
mangum = "^0.19.0"
geojson-pydantic = "^1.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
